Add fileContent method to Response extension
This commit adds a new feature to the MicroFramework's Response extension, a method called 'fileContents'. This new method reads the contents of a file and sends it as a response, also allowing the option for specifying the content type. Additionally, it will throw an exception if the file does not exist.