Add error handling method to View class
Incorporated a new static function called `renderErrorPage` to handle and display exceptions. This method not only takes in the throwable exception as input but also purges the output buffer to ensure that any previous parts of the page that had been rendered are discarded, thus clearly displaying the error page when exceptions occur.