Disable database connection and refactor error page
This commit disables the database connection in the .env settings. Furthermore, the structure of the error page's source code has been modified to be more simplified and straightforward. Changes to the error page include a switch to the Tailwind CSS framework to standardize design across the site and improve maintainability.