Add dialog box configuration to Controller and Ajax
A dialog box configuration has been added to the Controller class, enabling dynamic changes to the dialog box settings such as width. This configuration is then used in Ajax to adjust the dialog box display accordingly. The use of json encoding ensures the configuration can be easily accessed in Ajax. Another change includes the deletion of the `template.twig` file.